requirements.txt
<<<<<<< HEAD
python-telegram-bot==20.7
python-dotenv
=======
python-telegram-bot==20.5
python-dotenv==1.0.0
nltk==3.8.1
httpx==0.26.2
anyio==3.7.0
>>>>>>> 3c6e644 (Fixed spam deletion and improved moderation logic)
